requirements.txt
transformers==4.30.2
wandb
python-dotenv
bitsandbytes
deepseed